# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Clinic Launch OS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects/{projectId}/financial:
    get:
      operationId: getFinancialModel
      tags: [financials]
      summary: Get financial model for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Financial model
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinancialModel"
    put:
      operationId: upsertFinancialModel
      tags: [financials]
      summary: Create or update financial model
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertFinancialModelBody"
      responses:
        "200":
          description: Financial model
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinancialModel"
  /projects/{projectId}/financial/calculate:
    post:
      operationId: calculateFinancials
      tags: [financials]
      summary: Calculate financial outputs for a given scenario
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CalculateFinancialsBody"
      responses:
        "200":
          description: Calculated financial outputs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinancialCalculation"

  # Dashboard
  /projects/{projectId}/dashboard:
    get:
      operationId: getProjectDashboard
      tags: [dashboard]
      summary: Get dashboard summary for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
  /projects/{projectId}/risk-flags:
    get:
      operationId: getRiskFlags
      tags: [dashboard]
      summary: Get risk flags for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Risk flags
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RiskFlag"
...
  /projects/{projectId}/cashflow:
    get:
      operationId: getProjectCashflow
      tags: [financials]
      summary: Get monthly cashflow projection for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: scenario
          in: query
          required: false
          schema:
            type: string
            enum: [conservative, realistic, aggressive]
      responses:
        "200":
          description: Monthly cashflow data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CashflowMonth"
components:
  schemas:
...
    FinancialModel:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        rentGbp:
          type: number
        ratesGbp:
          type: number
        utilitiesGbp:
          type: number
        internetGbp:
          type: number
        insuranceGbp:
          type: number
        accountantGbp:
          type: number
        softwareGbp:
          type: number
        wasteContractGbp:
          type: number
        cleanerGbp:
          type: number
        subscriptionsGbp:
          type: number
        financeRepaymentsGbp:
          type: number
        stockPercent:
          type: number
        marketingGbp:
          type: number
        staffingGbp:
          type: number
        commissionsPercent:
          type: number
        consumablesGbp:
          type: number
        averageClientValueGbp:
          type: number
        treatmentRoomsCount:
          type: integer
        practitionerHoursPerDay:
          type: number
        workingDaysPerMonth:
          type: integer
        conservativeOccupancyPercent:
          type: number
        realisticOccupancyPercent:
          type: number
        aggressiveOccupancyPercent:
          type: number
        repeatBookingRatePercent:
          type: number
        membershipRevenueGbp:
          type: number
        existingClinicRevenueGbp:
          type: number
        ownerDrawingsGbp:
          type: number
        runwaySavingsGbp:
          type: number
        personalSalaryNeedsGbp:
          type: number
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - projectId
        - createdAt
        - updatedAt
    UpsertFinancialModelBody:
      type: object
      properties:
        rentGbp:
          type: number
        ratesGbp:
          type: number
        utilitiesGbp:
          type: number
        internetGbp:
          type: number
        insuranceGbp:
          type: number
        accountantGbp:
          type: number
        softwareGbp:
          type: number
        wasteContractGbp:
          type: number
        cleanerGbp:
          type: number
        subscriptionsGbp:
          type: number
        financeRepaymentsGbp:
          type: number
        stockPercent:
          type: number
        marketingGbp:
          type: number
        staffingGbp:
          type: number
        commissionsPercent:
          type: number
        consumablesGbp:
          type: number
        averageClientValueGbp:
          type: number
        treatmentRoomsCount:
          type: integer
        practitionerHoursPerDay:
          type: number
        workingDaysPerMonth:
          type: integer
        conservativeOccupancyPercent:
          type: number
        realisticOccupancyPercent:
          type: number
        aggressiveOccupancyPercent:
          type: number
        repeatBookingRatePercent:
          type: number
        membershipRevenueGbp:
          type: number
        existingClinicRevenueGbp:
          type: number
        ownerDrawingsGbp:
          type: number
        runwaySavingsGbp:
          type: number
        personalSalaryNeedsGbp:
          type: number
    CalculateFinancialsBody:
      type: object
      properties:
        scenario:
          type: string
          enum: [conservative, realistic, aggressive]
      required:
        - scenario
    FinancialCalculation:
      type: object
      properties:
        scenario:
          type: string
          enum: [conservative, realistic, aggressive]
        monthlyRevenue:
          type: number
        annualRevenue:
          type: number
        monthlyFixedCosts:
          type: number
        monthlyVariableCosts:
          type: number
        monthlyTotalCosts:
          type: number
        monthlyNetProfit:
          type: number
        annualNetProfit:
          type: number
        ebitda:
          type: number
        cashRunwayMonths:
          type: number
        breakEvenRevenueGbp:
          type: number
        breakEvenOccupancyPercent:
          type: number
        minimumViableRevenueGbp:
          type: number
        safeOperatingThresholdGbp:
          type: number
        occupancyUsedPercent:
          type: number
        monthsUntilProfitable:
          type: number
          nullable: true
      required:
        - scenario
        - monthlyRevenue
        - annualRevenue
        - monthlyFixedCosts
        - monthlyVariableCosts
        - monthlyTotalCosts
        - monthlyNetProfit
        - annualNetProfit
        - ebitda
        - cashRunwayMonths
        - breakEvenRevenueGbp
        - breakEvenOccupancyPercent
        - minimumViableRevenueGbp
        - safeOperatingThresholdGbp
        - occupancyUsedPercent
    DashboardSummary:
      type: object
      properties:
        projectId:
          type: integer
        projectName:
          type: string
        status:
          type: string
        targetOpeningDate:
          type: string
          nullable: true
        daysToOpening:
          type: integer
          nullable: true
        launchReadinessPercent:
          type: integer
        totalTaskCount:
          type: integer
        completedTaskCount:
          type: integer
        blockedTaskCount:
          type: integer
        highRiskTaskCount:
          type: integer
        criticalRiskFlagCount:
          type: integer
        totalProjectCostLow:
          type: number
        totalProjectCostMid:
          type: number
        totalProjectCostHigh:
          type: number
        currentSelectedCost:
          type: number
        projectedFirstYearProfit:
          type: number
          nullable: true
        monthlyBurnRate:
          type: number
          nullable: true
        cashRunwayMonths:
          type: number
          nullable: true
        projectConfidenceScore:
          type: integer
        phaseProgress:
          type: array
          items:
            $ref: "#/components/schemas/PhaseProgress"
      required:
        - projectId
        - projectName
        - status
        - launchReadinessPercent
        - totalTaskCount
        - completedTaskCount
        - blockedTaskCount
        - highRiskTaskCount
        - criticalRiskFlagCount
        - totalProjectCostLow
        - totalProjectCostMid
        - totalProjectCostHigh
        - currentSelectedCost
        - projectConfidenceScore
        - phaseProgress
    PhaseProgress:
      type: object
      properties:
        phaseId:
          type: integer
        phaseName:
          type: string
        completedTasks:
          type: integer
        totalTasks:
          type: integer
        percentComplete:
          type: integer
        status:
          type: string
      required:
        - phaseId
        - phaseName
        - completedTasks
        - totalTasks
        - percentComplete
        - status
...
    RiskFlag:
      type: object
      properties:
        level:
          type: string
          enum: [warning, critical]
        category:
          type: string
        message:
          type: string
        taskId:
          type: integer
          nullable: true
        taskTitle:
          type: string
          nullable: true
      required:
        - level
        - category
        - message
    CashflowMonth:
      type: object
      properties:
        month:
          type: integer
        monthLabel:
          type: string
        revenue:
          type: number
        fixedCosts:
          type: number
        variableCosts:
          type: number
        netCashflow:
          type: number
        cumulativeCashflow:
          type: number
        isBreakevenMonth:
          type: boolean
      required:
        - month
        - monthLabel
        - revenue
        - fixedCosts
        - variableCosts
        - netCashflow
        - cumulativeCashflow
        - isBreakevenMonth