# lib/api-zod/src/generated/types/cashflowMonth.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Clinic Launch OS API
 * OpenAPI spec version: 0.1.0
 */

export interface CashflowMonth {
  month: number;
  monthLabel: string;
  revenue: number;
  fixedCosts: number;
  variableCosts: number;
  netCashflow: number;
  cumulativeCashflow: number;
  isBreakevenMonth: boolean;
}