# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Clinic Launch OS API
 * OpenAPI spec version: 0.1.0
 */
import type { PhaseProgress } from "./phaseProgress";

export interface DashboardSummary {
  projectId: number;
  projectName: string;
  status: string;
  targetOpeningDate?: string | null;
  daysToOpening?: number | null;
  launchReadinessPercent: number;
  totalTaskCount: number;
  completedTaskCount: number;
  blockedTaskCount: number;
  highRiskTaskCount: number;
  criticalRiskFlagCount: number;
  totalProjectCostLow: number;
  totalProjectCostMid: number;
  totalProjectCostHigh: number;
  currentSelectedCost: number;
  projectedFirstYearProfit?: number | null;
  monthlyBurnRate?: number | null;
  cashRunwayMonths?: number | null;
  projectConfidenceScore: number;
  phaseProgress: PhaseProgress[];
}