# lib/api-zod/src/generated/types/phaseProgress.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Clinic Launch OS API
 * OpenAPI spec version: 0.1.0
 */

export interface PhaseProgress {
  phaseId: number;
  phaseName: string;
  completedTasks: number;
  totalTasks: number;
  percentComplete: number;
  status: string;
}